ref : ViewModelで記述していたHTTP周りの記述を別クラスに転記した。
